App loads from iso storage then downloads in background
Rather than every 2 hours, forcing the user to wait while the fresh
OData is downloaded before the app is usable. The app now loads the data
in Isolated storage every time. Then if it is time to refesh, the new
data is downloaded in the background, and the isolated storage is
upated.